+++ CHANGELOG.md
@@ -34,6 +34,7 @@
   `Series.to_snowpark()`, and `Series.to_snowflake()` on the "Pandas" and "Ray"
   backends. Previously, only some of these functions and methods were supported
   on the Pandas backend.
+- Added support for `Index.get_level_values()`.
 
 #### Improvements
 - Set the default transfer limit in hybrid execution for data leaving Snowflake to 100k, which can be overridden with the SnowflakePandasTransferThreshold environment variable. This configuration is appropriate for scenarios with two available engines, "Pandas" and "Snowflake" on relational workloads.